+++ utils/hexparse.h
@@ -5,6 +5,6 @@
 
 void write_raw_data_to_hex_buffer(unsigned char* raw_data, size_t raw_len, char* hex_buffer);
 int is_valid_hex_char(char c);
-int parse_hex_data(const char* hex_data, unsigned char* out_buff, int exp_len, int* out_len);
+int parse_hex_data(const char* hex_data, unsigned char* out_buff, int exp_len, size_t* out_len);
 
 #endif
